feat: update naming convention and folder structure
