+++ src/Appwrite/Utopia/Database/Validator/Query/Cursor.php
@@ -1,0 +1,44 @@
+<?php
+
+namespace Appwrite\Utopia\Database\Validator\Query;
+
+use Utopia\Validator;
+use Utopia\Database\Query;
+use Utopia\Database\Validator\UID;
+
+class Cursor extends Validator
+{
+    /**
+     * @var string
+     */
+    protected $message = 'Invalid query';
+
+    /**
+     * Is valid.
+     *
+     * Returns true if method is cursorBefore or cursorAfter and value is not null
+     *
+     * Otherwise, returns false
+     *
+     * @param Query $value
+     *
+     * @return bool
+     */
+    public function isValid($query): bool
+    {
+        // Validate method
+        $method = $query->getMethod();
+
+        if ($method === Query::TYPE_CURSORAFTER || $method === Query::TYPE_CURSORBEFORE) {
+            $cursor = $query->getValue();
+            $validator = new UID();
+            if ($validator->isValid($cursor)) {
+                return true;
+            }
+            $this->message = 'Invalid cursor: ' . $validator->getDescription();
+            return false;
+        }
+
+        return false;
+    }
+}
