+++ src/Appwrite/Utopia/Database/Validator/Query/Limit.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Appwrite\Utopia\Database\Validator\Queries;
+namespace Appwrite\Utopia\Database\Validator\Query;
 
 use Utopia\Database\Query;
 use Utopia\Validator\Range;
 use Utopia\Validator;
 
-class LimitOffsetQuery extends Validator
+class Limit extends Validator
 {
     /**
      * @var string
@@ -14,19 +14,15 @@ class LimitOffsetQuery extends Validator
     protected $message = 'Invalid query';
 
     protected int $maxLimit;
-    protected int $maxOffset;
 
     /**
      * Query constructor
      *
      * @param int $maxLimit
-     * @param int $maxOffset
-     * @param int $maxValuesCount
      */
-    public function __construct(int $maxLimit = 100, int $maxOffset = 5000)
+    public function __construct(int $maxLimit = 100)
     {
         $this->maxLimit = $maxLimit;
-        $this->maxOffset = $maxOffset;
     }
 
     /**
@@ -52,21 +48,10 @@ class LimitOffsetQuery extends Validator
         return false;
     }
 
-    protected function isValidOffset($offset): bool
-    {
-        $validator = new Range(0, $this->maxOffset);
-        if ($validator->isValid($offset)) {
-            return true;
-        }
-
-        $this->message = 'Invalid offset: ' . $validator->getDescription();
-        return false;
-    }
-
     /**
      * Is valid.
      *
-     * Returns true if method is limit or offset and values are within range.
+     * Returns true if method is limit values are within range.
      *
      * @param Query $value
      *
@@ -76,19 +61,14 @@ class LimitOffsetQuery extends Validator
     {
         // Validate method
         $method = $query->getMethod();
-        switch ($method) {
-            case Query::TYPE_LIMIT:
-                $limit = $query->getValue();
-                return $this->isValidLimit($limit);
 
-            case Query::TYPE_OFFSET:
-                $offset = $query->getValue();
-                return $this->isValidOffset($offset);
-
-            default:
-                $this->message = 'Query method invalid: ' . $method;
-                return false;
+        if ($method !== Query::LIMIT) {
+            $this->message = 'Query method invalid: ' . $method;
+            return false;
         }
+
+        $limit = $query->getValue();
+        return $this->isValidLimit($limit);
     }
 
     /**
